refactor: View 클래스를 InputView와 OutputView로 분리
- App.js에서 readLineLottoInputs 삭제
