+++ src/view/OutputView.js
@@ -1,18 +1,7 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
-import Validator from "./Validator.js";
-import LottoUtils from "./LottoUtils.js";
+import LottoUtils from "../LottoUtils.js";
 
-class View {
-  static async readLinePrice() {
-    const input = await MissionUtils.Console.readLineAsync('구입금액을 입력해 주세요.\n')
-    const inputPrice = parseInt(input, 10);
-
-    Validator.validatePrice(input);
-    Validator.validateParsePrice(inputPrice);
-
-    return inputPrice;
-  }
-
+class OutputView {
   static printLotto(count) {
     this.printNewLine();
     MissionUtils.Console.print(`${count}개를 구매했습니다.`);
@@ -28,29 +17,6 @@ class View {
     });
 
     return lottoArray;
-  }
-
-  static async readLineNumber() {
-    this.printNewLine();
-    const input = await MissionUtils.Console.readLineAsync('당첨 번호를 입력해 주세요.\n');
-    const trimInput = input.toString().trim().split(',');
-    const parseNumber = LottoUtils.getParsingNumber(trimInput);
-
-    Validator.validateNumber(trimInput);
-    Validator.validateParseNumber(parseNumber);
-
-    return parseNumber;
-  }
-
-  static async readLineBonusNumber() {
-    this.printNewLine();
-    const input = await MissionUtils.Console.readLineAsync('보너스 번호를 입력해 주세요.\n');
-    const parseBonusNumber = parseInt(input, 10);
-
-    Validator.validateBonusNumber(input);
-    Validator.validateParseBonusNumber(parseBonusNumber);
-
-    return parseBonusNumber;
   }
 
   static printNewLine() {
@@ -75,4 +41,4 @@ class View {
   }
 }
 
-export default View;
+export default OutputView;
